refactor: 認証関連ファイルをlib/feature/auth配下に移動
